requirements.txt
cryptography==38.0.1
paho-mqtt==1.6.1